librados: use uint64_t for extent len in extmap

+++ src/include/rados/librados.hpp
@@ -122,7 +122,7 @@ namespace librados
     int read(const std::string& oid, bufferlist& bl, size_t len, uint64_t off);
     int remove(const std::string& oid);
     int trunc(const std::string& oid, uint64_t size);
-    int mapext(const std::string& o, uint64_t off, size_t len, std::map<uint64_t, size_t>& m);
+    int mapext(const std::string& o, uint64_t off, size_t len, std::map<uint64_t,uint64_t>& m);
     int sparse_read(const std::string& o, std::map<uint64_t, size_t>& m, bufferlist& bl, size_t len, uint64_t off);
     int getxattr(const std::string& oid, const char *name, bufferlist& bl);
     int getxattrs(const std::string& oid, std::map<std::string, bufferlist>& attrset);
@@ -168,7 +168,7 @@ namespace librados
     int aio_read(const std::string& oid, AioCompletion *c,
 		 bufferlist *pbl, size_t len, uint64_t off);
     int aio_sparse_read(const std::string& oid, AioCompletion *c,
-			std::map<uint64_t,size_t> *m, bufferlist *data_bl,
+			std::map<uint64_t,uint64_t> *m, bufferlist *data_bl,
 			size_t len, uint64_t off);
     int aio_write(const std::string& oid, AioCompletion *c, const bufferlist& bl,
 		  size_t len, uint64_t off);
